Payment related bug fixes, ride confirmation page reworked and shifted feedback component
