+++ src/components/RideConfirmation/RideConfirmation.css
@@ -1,0 +1,81 @@
+/* Font */
+@import url('https://fonts.googleapis.com/css?family=Raleway:100,200,300,400,500,600,700,800,900&display=swap');
+
+/* General Styling */
+html, body {
+    font-family: 'Raleway', sans-serif;  
+}
+
+/* Header Styling */
+.thankyou-page ._header {
+    background: #34a853 url(https://codexcourier.com/images/main_page.jpg) center/cover no-repeat;
+    padding: 100px 30px;
+    text-align: center;
+}
+.thankyou-page ._header h1 {
+    font-size: 65px;
+    font-weight: 800;
+    color: white;
+    margin: 0;
+}
+
+/* Body Styling */
+.thankyou-page ._body ._box {
+    margin: auto;
+    max-width: 90%;
+    padding: 50px;
+    background: white;
+    border-radius: 3px;
+    box-shadow: 0 0 35px rgba(10, 10, 10, 0.12);
+    -moz-box-shadow: 0 0 35px rgba(10, 10, 10, 0.12);
+    -webkit-box-shadow: 0 0 35px rgba(10, 10, 10, 0.12);
+}
+.thankyou-page ._body ._box h2 {
+    font-size: 32px;
+    font-weight: 600;
+    color: #4ab74a;
+}
+.gap {
+    margin-top: 100px;
+}
+
+/* Media Queries */
+@media only screen and (min-width: 1224px) {
+    /* Header Styling */
+    .thankyou-page ._header .logo {
+        max-width: 200px;
+        margin: 0 auto 50px;
+    }
+    .thankyou-page ._header .logo img {
+        width: 100%;
+    }
+
+    /* Body Styling */
+    .thankyou-page ._body {
+        margin: -80px 0 30px;
+    }
+
+    /* Footer Styling */
+    .thankyou-page ._footer {
+        text-align: center;
+        padding: 50px 30px;
+    }
+    .thankyou-page ._footer .btn {
+        background: #4ab74a;
+        color: white;
+        border: 0;
+        font-size: 14px;
+        font-weight: 600;
+        border-radius: 0;
+        letter-spacing: 0.8px;
+        padding: 20px 33px;
+        text-transform: uppercase;
+    }
+}
+
+@media only screen and (max-width: 480px) {
+    /* Media-specific styling */
+    .media-thank-h2 {
+        text-align: left;
+    }
+}
